use `ZnInit` package
- remove automatic init code
- remove descriptor code
- adapt existing code to work with the `ZnInit` package
