+++ pyproject.toml
@@ -17,6 +17,7 @@ dot4dict = "^0.1.1"
 tqdm = "^4.64.0"
 pandas = "^1.4.3"
 pygit2 = ">=1.9,<1.11"
+zninit = "^0.1.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.1.2"
