+++ zntrack/zn/nodes.py
@@ -31,14 +31,14 @@ class Nodes(ZnTrackOption):
     zn_type = utils.ZnTypes.DEPS
     file = utils.Files.zntrack
 
-    def __init__(self, default_value=None, **kwargs):
-        if default_value is not None:
+    def __init__(self, default=None, **kwargs):
+        if default is not None:
             raise ValueError(
                 "zn.Nodes does not support default values because they can be mutable."
                 " Please set the default value in the __init__"
             )
 
-        super().__init__(**kwargs)
+        super().__init__(default=default, **kwargs)
 
     def __set__(self, instance, value):
         """Include type check for better error reporting"""
